refactor(components/ui/form): keep default patterns
