+++ src/presentation/components/ui/form/index.ts
@@ -7,7 +7,7 @@ import { Item } from './item'
 import { Label } from './label'
 import { Message } from './message'
 
-export { useFormField } from './hooks/useFormField'
+export { useFormField, useHookForm } from './hooks'
 export const Form = {
 	Provider: FormProvider,
 	Control,
