+++ src/presentation/components/ui/form/item.tsx
@@ -1,27 +1,26 @@
-import React, { useId, useMemo } from 'react'
+import { HTMLAttributes, forwardRef, useId, useMemo } from 'react'
 
 import { cn } from '@/main/utils'
 
 import { FormItemContext } from './contexts'
 
-export const Item = React.forwardRef<
-	HTMLDivElement,
-	React.HTMLAttributes<HTMLDivElement>
->(({ className, ...props }, ref) => {
-	const id = useId()
+export const Item = forwardRef<HTMLDivElement, HTMLAttributes<HTMLDivElement>>(
+	({ className, ...props }, ref) => {
+		const id = useId()
 
-	const providerProps = useMemo(
-		() => ({
-			id
-		}),
-		[id]
-	)
+		const providerProps = useMemo(
+			() => ({
+				id
+			}),
+			[id]
+		)
 
-	return (
-		<FormItemContext.Provider value={providerProps}>
-			<div ref={ref} className={cn('space-y-2', className)} {...props} />
-		</FormItemContext.Provider>
-	)
-})
+		return (
+			<FormItemContext.Provider value={providerProps}>
+				<div ref={ref} className={cn('space-y-2', className)} {...props} />
+			</FormItemContext.Provider>
+		)
+	}
+)
 
 Item.displayName = 'FormItem'
