+++ src/presentation/components/ui/form/label.tsx
@@ -1,16 +1,15 @@
-import React from 'react'
+import { forwardRef, ElementRef, ComponentPropsWithoutRef } from 'react'
 
-import * as LabelPrimitive from '@radix-ui/react-label'
+import { Root } from '@radix-ui/react-label'
 
 import { cn } from '@/main/utils'
-
-import { Label as AppLabel } from '../label'
+import { Label as AppLabel } from '@/presentation/components/ui/label'
 
 import { useFormField } from './hooks/useFormField'
 
-export const Label = React.forwardRef<
-	React.ElementRef<typeof LabelPrimitive.Root>,
-	React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root>
+export const Label = forwardRef<
+	ElementRef<typeof Root>,
+	ComponentPropsWithoutRef<typeof Root>
 >(({ className, ...props }, ref) => {
 	const { formItemId } = useFormField()
